OpenXava/Spring integration code refactored. Datasource integration finished
